Notificaciones a servidor

+++ TThrough/mvvm/ViewModel/PopUpSolicitudesPendientesViewModel.cs
@@ -1,33 +1,38 @@
 ﻿using Microsoft.EntityFrameworkCore;
 using System;
 using System.Collections.Generic;
+using System.Collections.ObjectModel;
 using System.Linq;
 using System.Text;
+using System.Text.Json;
 using System.Threading.Tasks;
 using System.Windows.Input;
 using TThrough.data;
 using TThrough.Entidades;
+using TThrough.mvvm.Models;
 
 namespace TThrough.mvvm.ViewModel
 {
     public class PopUpSolicitudesPendientesViewModel : ViewModelBase
     {
         #region Propiedades
+        private Servicios.ServicioTCP tcp;
+
         private readonly TalkthroughContext _context;
 
         private Models.Usuario _usuarioConectado;
 
-        public Action<Models.Chats> ChatCreado {  get; set; }
+        public Action<Models.Chats> ChatCreado { get; set; }
 
         public ICommand añadirUser => new RelayCommand(
-            _=> AceptarSolicitud(),
-            _=> true);
+            _ => AceptarSolicitud(),
+            _ => true);
 
         public ICommand rechazarUser => new RelayCommand(
             _ => RechazarSolicitud(),
             _ => true);
 
-        public List<VistaSolicitudes> Solicitudes { get; set; }
+        public ObservableCollection<VistaSolicitudes> Solicitudes { get; set; }
 
         private VistaSolicitudes _solicitudSeleccionada { get; set; }
 
@@ -43,12 +48,13 @@ namespace TThrough.mvvm.ViewModel
 
 
         #region Constructores
-        public PopUpSolicitudesPendientesViewModel(Models.Usuario u)
+        public PopUpSolicitudesPendientesViewModel(Models.Usuario u, Servicios.ServicioTCP tcp)
         {
-            Solicitudes = new List<VistaSolicitudes>();
+            Solicitudes = new();
             _context = TalkthroughContextFactory.SendContextFactory();
             _usuarioConectado = u;
             CargarSolicitudes();
+            this.tcp = tcp;
         }
         #endregion
 
@@ -73,73 +79,84 @@ namespace TThrough.mvvm.ViewModel
 
         private void RechazarSolicitud()
         {
-            var usuario = _context.Usuarios.Single(x => x.NombreUsuario == SolicitudSeleccionada.NombreUsuario);
-
-            if (usuario != null)
+            if (SolicitudSeleccionada != null)
             {
-                var solicitud = _context.Amigos.Single(x => x.IdUsuarioEnvio == usuario.IdUsuario);
-                _context.Amigos.Remove(solicitud);
-                _context.SaveChanges();
-            }
+                var usuario = _context.Usuarios.SingleOrDefault(x => x.NombreUsuario == SolicitudSeleccionada.NombreUsuario);
 
-            Solicitudes.Clear();
-            CerrarPopupAction?.Invoke();
+                if (usuario != null)
+                {
+                    var solicitud = _context.Amigos.SingleOrDefault(x => x.IdUsuarioEnvio == usuario.IdUsuario && x.IdUsuarioRemitente == _usuarioConectado.IdUsuario);
+                    if (solicitud != null)
+                    {
+                        _context.Amigos.Remove(solicitud);
+                        _context.SaveChanges();
+                    }
+
+                    Solicitudes.Remove(SolicitudSeleccionada);
+                    if (!Solicitudes.Any())
+                        CerrarPopupAction?.Invoke();
+                }
+            }
         }
         #endregion
 
-        private void AceptarSolicitud() 
+        private void AceptarSolicitud()
         {
+            var usuario = _context.Usuarios.SingleOrDefault(x => x.NombreUsuario == SolicitudSeleccionada.NombreUsuario);
+            if (usuario == null) return;
 
-            //Recoge el usuario que manda la solicitud al cliente que está aceptando solicitudes.
-            var usuario = _context.Usuarios.Single(x => x.NombreUsuario == SolicitudSeleccionada.NombreUsuario);
-
-            //Modifica el estado de la solicitud a verdadero.
-            if (usuario != null) 
+            var solicitud = _context.Amigos.SingleOrDefault(x => x.IdUsuarioEnvio == usuario.IdUsuario && x.IdUsuarioRemitente == _usuarioConectado.IdUsuario);
+            if (solicitud != null)
             {
-                
-                var solicitud = _context.Amigos.Single(x=>x.IdUsuarioEnvio ==usuario.IdUsuario && x.IdUsuarioRemitente==_usuarioConectado.IdUsuario);
-
                 solicitud.SolicitudAceptada = true;
-
                 _context.Amigos.Update(solicitud);
-               
             }
 
-
-            //Crea un nuevo chat con el nombre del amigo
             var nuevoChat = new Models.Chats
             {
                 IdChat = Guid.NewGuid().ToString(),
-                NombreChat = usuario.NombrePublico,
+                NombreChat = usuario.NombreUsuario,
                 FotoChat = usuario.FotoPerfil,
                 FechaInicioChat = DateTime.Now,
             };
+
+
             _context.Chats.Add(nuevoChat);
 
-            //Crea una entrada en la tabla intermedia de manera que relacione a los dos usuarios con el chat
-            _context.ChatsUsuarios.Add(new Models.ChatsUsuarios
-            {
-                IdChat = nuevoChat.IdChat,
-                IdUsuario = usuario.IdUsuario 
-            });
-
-            _context.ChatsUsuarios.Add(new Models.ChatsUsuarios
-            {
-                IdChat = nuevoChat.IdChat,
-                IdUsuario = _usuarioConectado.IdUsuario 
-            });
+            _context.ChatsUsuarios.Add(new Models.ChatsUsuarios { IdChat = nuevoChat.IdChat, IdUsuario = usuario.IdUsuario });
+            _context.ChatsUsuarios.Add(new Models.ChatsUsuarios { IdChat = nuevoChat.IdChat, IdUsuario = _usuarioConectado.IdUsuario });
 
             _context.SaveChanges();
-
-            
             ChatCreado?.Invoke(nuevoChat);
-            Solicitudes.Clear();
-           
-            CerrarPopupAction?.Invoke();
+            AceptarSolicitud(usuario,nuevoChat);
+            Solicitudes.Remove(SolicitudSeleccionada);
+            if (!Solicitudes.Any())
+                CerrarPopupAction?.Invoke();
         }
 
-        
-    
+        private async void AceptarSolicitud(Usuario usuarioSeleccionado, Chats chat)
+        {
+
+            var chatEnviar = new Chats
+            {
+                IdChat = chat.IdChat,
+                NombreChat = usuarioSeleccionado.NombreUsuario,
+                FechaInicioChat = chat.FechaInicioChat,
+            };
+
+            var mensaje = new MensajeJson
+            {
+                Tipo = "solicitud_aceptada",
+                Emisor = _usuarioConectado.IdUsuario,
+                Receptor = usuarioSeleccionado.IdUsuario,
+                Datos = chatEnviar,
+            };
+
+            var json = JsonSerializer.Serialize(mensaje);
+
+            await tcp.EnviarMensaje(json);
+        }
+
     }
-    
+
 }
